[DEV-3801] changes after review
