+++ package.json
@@ -32,6 +32,7 @@
     "@deuro/eurocoin": "^1.0.12",
     "@dhedge/v2-sdk": "^1.9.8",
     "@eth-optimism/sdk": "^3.2.0",
+    "@frankencoin/zchf": "^0.2.18",
     "@maticnetwork/maticjs": "^3.7.8",
     "@maticnetwork/maticjs-ethers": "^1.0.3",
     "@nestjs-modules/mailer": "^1.10.3",
